backend/requirements.txt
fastapi==0.105.0
python-multipart==0.0.6
msoffcrypto-tool==5.1.1
uvicorn==0.24.0